Stop distingishing functions vs. bindings + merge case and `<-` behaviors
Additionally, dissolve BindGroup and map the symbols that it contained
immediately to/from the bindee (the thing being bound).

This commit does a lot, generally simplifying the logic of `pt_search` so
it doesn't have to treat the bindings of "functions" (i.e. things bound with
`=`) from pattern binds (i.e. things bound with `<-` or `case`). With this
additional clarity, we can correctly point symbols to their dependents. The
new scheme is much simpler to explain: the point store (PtStore) consists of
applications and bindings that ping-pong between each other (all symbols in
apps are dependent on the binders at their binding sites; all binders are
dependent on the symbols in the app groups they're found in).

Additionally, the handling of case and `<-` are mostly unified since they share
so much in common.