Thread "visited" as state through pt_search + deduplicate idents
Allay fears that the program is non-converging by getting rid of a lot of
redundant work:

1. Stop splitting `visited` across each branch of the `pt_search` DFS and
instead thread it with a State monad;
2. Deduplicate the symbols in each level. Lots of symbols are used more than
once in app groups, sometimes way more than once like `.`.

This allowed the traversal to halt tracking a symbol in stacker itself, which
is a start. Yet to test other cases.