Fix where `mk_pt_store` looks for fn names & args in function bindings
The previous version of the handler for function types just looked for
the "FnBind" site, which failed for functions with multiple matches.
Note the HIE AST collapses singleton lists of things into the same node
as its parent, hence why many constructors end up at each node. Before,
we couldn't find a function name because it's really the Match that's 1-1
with the function name and args.