Correct dependency rules between binds and apps
Several of the rules for dependency were misjudged and needed to be
corrected. The right way to traverse the dependency graph is:

1. From an app group:
   1. Look to see if anything points directly to an argument in scope.
      If it does, then trace the argument to its parent function and
      look for app groups where that function is used + its RHS deps
      that it closes over.
   2. If it points instead to a symbol that's bound in scope, find the
      binding and collect all the RHS [shallow] app groups. Feed those
      into the recursion.
2. From a bind key:
   1. If it's a named binding, find its use sites with a map from names
      to app groups.
   2. If it's a lambda, find its use site with a segment tree from
      spans to app groups.

Furthermore, instead of passing groups of names that are "local" to an
AST in `mk_pt_store`, pass up AppGroups so they can be marked in the
graph in behavior 1.2.

Most of these changes manifested in changes to the Bind* datatypes in
the Lang, leading to changes in `mk_pt_store`. Additionally, the
`pt_search` recursion was changed substantially to obey the corrected
traversal behavior.