Traverse app groups in RHS of function + log all to visited
As I claimed a few commits ago but forgot to actually implement: the correct
behavior digging from a binding site is to look both up (for callsites) and
inwards (for values in the closure from the environment).

While we needed new code to log the app groups involved in these traversals,
I had never actually logged the ones in a separate case, from app groups
straight to other app groups. This commit also patches that hole.