Correct the edge directions for the dependency graph
Point `this_nodes` -> `next_nodes`, correcting a long-standing mistake
that tended to be covered up at small examples because bindings were
reflexively bound to their RHS app groups.

In the course of debugging, I've verified (visually) that mk_pt_store
is sane, binding the right arguments to functions and functions to RHS
app groups. It just needs to filter out datatypes that for the moment
it thinks are "bound" in the bindee positions.

Also verified visually that pt_search doesn't seem to be making any
transitive edges. All the edges look to be linking results from exactly
one step in the recursion, which is what we want: max granularity.