Refine criteria for identifying function name
Rather than precluding Case and LamCase from the AST stack, look to see if
a FunBind comes before a GRHS in the stack. Filtering by exclusion of
forbidden tags is a bit fragile, and the way it was implemented previously
was just plain wrong (looking up the whole stack).

Additionally, guard against when we can't find a function name. It
happened when scanning some code with MagicHash in `primitives`'s
Data.Primitives.Types which had some weird unboxed values that presented
in the AST as a bare FunBind with no body.