Fix dependency assignments by pairing idents with children
Previously, there was some overzealous concatting with the `to` and `from`
lists of idents during the `pt_search`: i.e. all of the idents at the
current iteration would be bound to all the children. Here, we pair all
of the idents to the children they produce, so they can be paired properly.
This required a bit of new StateT stuff to thread the `visited` set
through all the pairings.

In the meantime we refactor the graph handling so the graph is constructed
at the very end to save on perf, since we're really just producing edges.
(Eventually that will be a proper edge set/Vector for fast concat).