Add edge labels with "witnesses" to a dependency
Previously, we were losing information over the relationship between dependent
nodes in the dependency graph, e.g. an AppGroup was dependent on another
through a binding, but this binding wasn't recorded. Same with args and
BindKeys. Now, these missing symbols are added as witnesses in edge labels
so the graph is easier to understand.