Reduce lossiness & redundancy of graph type by refining edge labels & key type
Make edges a sum type and record all the identifiers along the path to get
from one node to another (e.g. the arg involved in an App -> Bind edge).
Adjacency list type has been ported to a plain IntMap instead of spans,
to reduce redundancy, since the node label on the righthand side should
have all the info we need. The UI won't allow arbitrary ranged selections
and even if they did, we could just rebuild the map on the viewer side