Include cost center ID in graph output
A surprisingly convoluted set of changes just to include the cost center
ID in the node type. That's because up to this point I'd always assumed
the only node info was the NodeKey. Now I had to generalize the type
everywhere to allow for a non-indexing piece of node info.