Handle `where`-clause bindings and pattern bindings properly
Sometimes under a `FunBind` there is the function binding as well as its
`where` clause on the same level. It's therefore necessary to stop the
depth iteration searching for idents in `arg_idents` at `HsValBinds`,
(which also required drilling down to avoid infinite loops when arg_idents
were called on an `HsValBinds`).

Pattern bindings that weren't part of a multi-match binding weren't
captured. They're a bit trickier than just adding `PatBind` to the list
of LHS tags because pattern bindings have such different structure to
function bindings: where the leftmost name of a function binding is the
only bindee of the GRHSs and the arguments are excluded, in the pattern
bind all names are bindees. Therefore the "arguments" coming from
arg_ident are in these cases pushed into the binding engine as function
names.