Use EdgeLabel from server + rethink data flow between components
This commit features a redesign of the interfaces between components so
that more of the information about snippets is preserved from the graph
definition down to the <Snip /> component and back up through the callbacks,
while preserving the encapsulation. In particular, a new "snip type" is
introduced to inform the rendering and click handlers in <CodeBlock />.

Srolling to snippets is also fixed.

Unfortunately, highlighted snippets don't fire events because highlight.js
colors the code block via `innerHTML`, so we need to get more creative with
the highlighting strategy. Probably going to use the label tree from hljs (
see hljs#2404 and hljs#1086)and just render the tree manually after
interspersing the snippet tags.