Generate highlight tree from hljs raw label tree through React
As mentioned in a previous commit, it's necessary to manually place the tags
(and split them as necessary to preserve the snippets) because hljs doesn't
preserve the tags present in the root node already.
`CodeBlock::render_highlight` dropped to recursively slice the highlight tree
up at the snippet points and nest them all properly.

Side issue is being explored about why some nodes with edges seem to be
filtered out by `gr2adjlist`.