Map filenames in spans to a file map index
Necessary space-saving measure now that we're going to multi-file things.
To that point, graph merging was also treating keys poorly, merging things
with the same key which didn't actually make any sense. What I really wanted
was to merge things with the same NodeKey, which will come as a later
optimization - for now, there may be redundancy in the final graph, which is
fine.

Incidentally, the function to replace all the span filenames with filemap idxs
 was too annoying to write without lenses, so Spand and friends were lens'd.