Fix graph edge assignment
I thought [incorrectly] that the edge labels at a given iteration were the
labels for the new edges. Really, they're the labels for the edges in the
_caller_. If you want, the node-label pairs correspond to a "-O" structure
that goes deeper to the right. This also matches the interface with the
base case, which takes an EdgeLabel to show the context for the initial
match.