Improve efficiency of interval flattening
Interval flattening is a crucial step in rendering intervals, where the
overlapping spans are all mutually divided, and the spans in theresulting
disjoint partition have all the keys of the spans that overlap there.
Before, I used an interval tree to search the endpoints of each span
and divide each set of overlapping intervals. However, this is very
expensive since for each iteration I need to search a growing number of
tags, plus the mutual division is n^2-ish. Some of the nodes have lots
of spans (like really big flat AppGroups) and this was destroying the
performance even testing small examples, so we needed to optimize this
now.

This introduces a simpler state machine that walks the spans in ascending
order of start character, and closes the tags off as the ends come in. It
has already proven loads more efficient in every test case so far.