Disambiguate spans of repeated symbols + introduce RevBindEdge
Previously, with LIdentifier as a key type for the symbol maps (and its
Ord instance ignoring the Span), the merging operation was assuming that
all the spans were the same anyways, which is totally wrong. Repeated calls
in the same AppGroup (ag_ident_map) and different binding matches
(bg_bnd_app_map) were losing the distinction in the bind information.
For now, this is just to mark these as separate clickable things to the same
node. In the future, different binding matches may well be separated by
moving the Span to the key type.

Along this line of thought, I realized that bindees' dependencies on their
rhs app groups fell back to one of the match groups, not the one
corresponding to the app group, and generally not a "canonical" one like
the first or biggest match. To highlight all the bind sites, I introduce
a "RevBindEdge" edge label to depend the bindee name on each match
separately.