Preview snippets in node/context lists
For improved UI, add a code block showing the surrounding code of context
and node snippets.

This was surprisingly complicated, partly because of encapsulation (where
I only want to render the hljs result once, preferably in CodeBlock, yet
these snippets need highlighted blocks) and partly because of the new
parse tree slicing that needed to be implemented to take advantage of the
pre-computed parse tree + make the syntax highlighting correct for
partial strings post-slicing.