Connect highlighting of "next nodes" list and snippets + fix scroll
A pair of seemingly simple fixes with plenty of complications, mostly due
to:

1. Avoiding re-renders with highlighting and scrolling, which used to
be controlled by the Snips. Now that refs are passed up to a parent
(MainContent), we have to a) keep things pure to keep the state stashing
of refs from forcing a re-render, and b) make refs a derived state of
the changes to the parsetree.
2. Double-click functionality on the list of next nodes is the best UI
because it makes the layout less dangerous to mouse around in vs.
hover to scroll to next position, and also allows a preview mode for the
next hop where we can toggle between the source and drain of the edge.
However, thanks to some quirks of DOM events
(https://stackoverflow.com/questions/64057672) clicking triggers both a
focus and click event, meaning it's possible to single-click and look like
a double-click. Besides listening for `tab` explicitly as a source of
focus (probably equally susceptible to races and out-of-ordering) the
best fix for now seems to be a debounce, albeit one that limits the
double-click speed which also isn't ideal. The dbl-click handler seems to
fill the gap without problems.
3. Purity of MainContent (to avoid needless re-highlights) meant that where
I wanted MainController to mostly concern itself with, well, control, it
still needed to render the context bar and tried to reach in and get a
parsetree from MainContent. This made MainContent logically impure and made
it miss some updates. So now MainController is responsible for even more,
now it also renders the syntax tree. I don't love this change -- calls into
question the advantage of making MainContent a PureComponent in the first
place, since CodeBlock has taken on the new important purity for perf.