Rollback moving view responsibilities to MainController
Yeah, I really wasn't feeling the change from 27ce5c2ea that moved the
responsibiliy of the view rendering up from MainContent to MainController
given the only change for functionality that was actually needed was to
make MainContent a Component. There are a few extra unecessary re-renders
this way due to the impurity of rendering the side context bar, but the
concerns seem better separated. The main hole is the only reason that
MainContent is impure: that the side context bars want code snippets
that have already been rendered once. This is either an acceptably
localized hole, or we could move hljs rendering up to the MainController
(which still breaks the encapsulation a bit).