Land new feature: toggling to go between two locations on edge
All the current edge types have two relevant points, usually a callsite
and a use-site. For example, an AppEdge has a binding and a given callsite
for that binding. It's very helpful to see both when deciding if we want
to go down an edge.

This change touches a lot of parts because it needs to stash an array of
toggleable snips, i.e. changing the type of soft_selected.