Implement multi-file and file tabs
Handle edges crossing a file boundary. There is some intricacy to doing
this, because toggling can change files, and we must be sure that the
downstream changes from setting that state don't try to change the file
back. We also want tabbing to a file to jump to the latest `at_idx`
position of that file. For two files at least, so far this works.