Refactor argument to CtxSnips and refine parsetree fns and types
It turns out that the parsetree type and TPreview were essentially the
same with different choices made for what payload to include and how
to deal with undefined. By merging them, the logic of handling parsetrees
becomes more common between `CtxSnip` and `CodeBlock`. The differences
that did exist however required lots of generalization for almost every
type related to ParseTree and most of the methods that handle them. In
particular, the difference between nullable SpanKey payload and nullable
SpanKey that was the source of confusion.