Standardize SnipItem (previously CtxSnip) rendering
Now, all of the context items use a common interface for rendering snips
via `spks2snips`, which converts MainSpanKeys into snippets along with
custom names. This tests the interface a little bit, since SpanKeys are
now also the primary rendering object for snippets.

It's a little clunky used in this way and could benefit from a renderable
interface that includes only the necessities, and it could also benefit
from stashing NodeKey somewhere. It just happens to benefit the "next
nodes" case and is tolerable for "history" and "this node".