Implement "show all" for ctx snippets
Wrap all the context lists in a component that counts events and reflects
them back into the renderer. Shame that such a simple stateful thing
needs to be so verbose, but unfortunately we can't have a prop-less
nameless hook ([I tried](https://jsfiddle.net/2y35nkgu/)) so this is the
best we've got for now.